fix clippy

+++ crates/service/src/index/mod.rs
@@ -240,8 +240,8 @@ impl<S: G> Index<S> {
             growing: view.growing.values().map(|x| x.len()).collect(),
             write: view.write.as_ref().map(|(_, x)| x.len()).unwrap_or(0),
             options: self.options().clone(),
-            size: view.growing.values().map(|x| x.size() as u64).sum::<u64>()
-                + view.sealed.values().map(|x| x.size() as u64).sum::<u64>(),
+            size: view.growing.values().map(|x| x.size()).sum::<u64>()
+                + view.sealed.values().map(|x| x.size()).sum::<u64>(),
         }
     }
 }
